Feature:
-Dynamic styles are created for the new releases section in the form of a dynamic carousel in position, size and opacity.
-Dynamic gif images are added to the categories section
