+++ src/stylesheets/Categories.css
@@ -18,6 +18,7 @@
   gap: 20px;
   align-items: center;
   justify-content: center;
+  cursor: pointer;
 }
 
 .categorie-card {
@@ -36,7 +37,10 @@
 }
 
 .img-container {
-  height: 85%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 80%;
 }
 
 .cat-img {
@@ -46,12 +50,17 @@
 
 .categorie-name {
   background-color: #000229;
-  color: #00f3ff;
+  color: #456dba;
   width: 100%;
   height: 20%;
   text-align: center;
-  border-bottom-left-radius: 10px;
-  border-bottom-right-radius: 10px;
+  font-size: 3.1vh;
+  border-bottom-left-radius: 30px;
+  border-bottom-right-radius: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-family: "Paytone One", sans-serif;
 }
 
 .Categories-Container {
@@ -62,7 +71,7 @@
   align-items: center;
   justify-content: center;
   background-color: #000229;
-  color: #00f3ff;
+  color: #153fd7;
   border-radius: 10px;
 }
