+++ src/stylesheets/gamereleases.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   align-items: center;
-  width: 65%;
+  width: 60%;
   height: 100%;
   border-left: 5px #00f3ff solid;
 }
@@ -29,7 +29,7 @@
 .releases {
   display: flex;
   flex-direction: row;
-  height: 500px;
+  height: 480px;
   width: 80%;
   justify-content: center;
   align-items: center;
@@ -38,14 +38,16 @@
 .release-card {
   display: flex;
   flex-direction: column;
-  width: 140px;
-  height: 280px;
-  border-radius: 10px;
-  background-color: #00f3ff;
   justify-content: center;
   align-items: center;
-  padding: 0.5%;
+  width: 140px;
+  height: 250px;
+  border-radius: 10px;
+  background-color: #01012b;
   gap: 2%;
+  transition: transform 0.5s ease;
+  position: absolute;
+  border: 3px #456dba solid;
 }
 
 .card-container {
@@ -60,17 +62,29 @@
 }
 
 .game-image {
-  width: 95%;
-  height: 95%;
+  width: 88%;
+  height: 70%;
+  border-radius: 12px;
 }
 
 .new {
   width: 100%;
-  height: 50px;
-  background-color: #01012b;
-  color: #00f3ff;
+  height: 20%;
+  background-color: #456dba;
+  color: #01012b;
   text-align: center;
+  height: max-content;
   margin: 0;
+  font-size: medium;
+  font-family: "Paytone One", sans-serif;
+}
+
+.platforms {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  height: 10%;
+  gap: 8%;
 }
 
 .button-container {
@@ -84,6 +98,9 @@
   /* top right bottom left */
   border-style: solid;
   border-color: #456dba;
+  border-radius: 10px;
+  height: 50px;
+  margin-bottom: 2%;
 }
 
 .side-button {
@@ -99,25 +116,51 @@
   width: 100%;
 }
 
-.release-card.main {
+.release-card.centered {
   transform: scale(1.2);
-  width: 175px;
   background-color: #000229;
   z-index: 100;
   color: #00f3ff;
-  border: 5px #00f3ff solid;
+  border: 3px #456dba solid;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  width: 140px;
+  height: 250px;
+  border-radius: 10px;
+  justify-content: center;
+  align-items: center;
 }
 
-.release-card.side {
-  transform: scale(1);
-  opacity: 0.4;
-  filter: blur(1.5px);
-  z-index: 70;
-}
-
-.release-card.side2 {
-  transform: scale(1.1);
-  opacity: 0.8;
-  filter: blur(1.2px);
+.release-card.left {
+  transform: translateX(-70%) scale(0.9);
   z-index: 80;
+  opacity: 1;
+  filter: blur(2px);
+}
+
+.release-card.right {
+  transform: translateX(70%) scale(0.9);
+  z-index: 80;
+  opacity: 1;
+  filter: blur(1.5px);
+}
+
+.release-card.left-small {
+  transform: translateX(-110%) scale(0.8);
+  z-index: 70;
+  opacity: 0.8;
+  filter: blur(2px);
+}
+
+.release-card.right-small {
+  transform: translateX(110%) scale(0.8);
+  z-index: 70;
+  opacity: 0.8;
+  filter: blur(2px);
+}
+
+.platforms-img {
+  height: 25px;
+  width: 25px;
 }
