+++ src/stylesheets/gamestop.css
@@ -3,7 +3,7 @@
 }
 
 .Ranking {
-  width: 35%;
+  width: 40%;
   height: 100%;
 }
 
@@ -15,22 +15,28 @@
   background-color: #01012b;
   color: #00f3ff;
   margin: 0;
-  height: 12%;
+  height: 10%;
   text-align: center;
   border: 5px #00f3ff solid;
   font-family: "Paytone One", sans-serif;
   font-weight: 400;
-  font-size: 4vh;
+  font-size: 3.5vh;
 }
 
 .list-top-container {
+  display: flex;
+  flex-direction: column;
   list-style: none;
   padding: 0;
   margin: 0;
-  border-width: 0 5px 0 5px; /* top right bottom left */
+  border-width: 0 5px 0 5px;
+  /* top right bottom left */
   border-style: solid;
   border-color: #00f3ff;
-  height: 87%;
+  height: 88%;
+  overflow-y: scroll;
+  gap: 2%;
+  background-color: #16397d;
 }
 
 .top-list {
@@ -39,27 +45,73 @@
   justify-content: space-between;
   align-items: center;
   margin: 0;
-  height: 10%;
-  padding: 0 3% 0 3%;
+  height: 30%;
+  padding: 0 1% 0 1%;
+  background-color: #01012b;
+}
+
+.top-title-container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  height: 35px;
+  background-color: #001b51;
+  width: 100%;
+  border-bottom-right-radius: 100px;
 }
 
 .list-item {
   padding: 0;
   margin: 0;
-  color: #00f3ff;
+  color: #6398FF;
+  font-size: larger;
+  font-family: "Paytone One", sans-serif;
+  font-weight: 400;
+  text-align: start;
 }
 
-.list-item-r {
+.top-info-container {
+  width: 74%;
+  height: 100%;
+  padding-left: 2%;
+}
+
+.game-rating-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 15%;
+  height: 90%;
   padding: 0;
   margin: 0;
-  color: #00f3ff;
   font-size: larger;
 }
 
-.top-list:nth-child(odd) {
-  background-color: #001b51;
+.list-item-r {
+  width: 80%;
+  height: 80%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  background-color: #8633f0;
+  padding: 0;
+  margin: 0;
+  color: #fff;
+  font-size: larger;
+  font-family: "Paytone One", sans-serif;
+  border-radius: 5px;
 }
 
-.top-list:nth-child(even) {
-  background-color: #1f3154;
+.top-image {
+  height: 95%;
+  width: 11%;
+}
+
+.game-date {
+  color: #fff;
+  font-family: "Paytone One", sans-serif;
+  font-size: medium;
+  text-align: start;
+  padding-top: 2%;
 }
