--- src/stylesheets/test.css
@@ -1,82 +1,0 @@
-.t-container {
-  height: 90%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.cards-tittle-container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 750px;
-  align-items: center;
-  justify-content: center;
-}
-
-.cards-section {
-  position: relative;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-}
-
-.cardt {
-  height: 480px;
-  width: 360px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  transition: transform 0.5s ease;
-  position: absolute;
-}
-
-.card-img {
-  height: 96%;
-  width: 96%;
-}
-
-.side-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-}
-
-.side-button:hover .button-img {
-  filter: brightness(150%);
-}
-
-.cardt.centered {
-  transform: scale(1.2);
-  background-color: #000229;
-  z-index: 100;
-  width: 340px;
-  color: #00f3ff;
-  border: 5px #00f3ff solid;
-  border-radius: 10px;
-}
-
-.cardt.left {
-  transform: translateX(-60%) scale(0.9);
-  z-index: 80;
-}
-
-.cardt.right {
-  transform: translateX(60%) scale(0.9);
-  z-index: 80;
-}
-
-.cardt.left-small {
-  transform: translateX(-100%) scale(0.8);
-  z-index: 70;
-}
-
-.cardt.right-small {
-  transform: translateX(100%) scale(0.8);
-  z-index: 70;
-}
